% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstparts-package.r, R/fstparts.R
\docType{package}
\name{fstparts}
\alias{fstparts}
\alias{fstparts-package}
\title{fstparts.}
\usage{
fstparts(data, dir = NULL, name = deparse(substitute(data)),
  overwrite = FALSE, compress = 0, ...)
}
\arguments{
\item{data}{data to be added to parts}

\item{dir}{directory where to create}

\item{name}{name of data file}

\item{overwrite}{\code{logical}: should a previous directory be removed?}

\item{compress}{should the data be compressed (passed through to \code{\link[fst:fst.write]{fst::fst.write()}})}
}
\description{
Large fst storage
}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com}

}
