% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chunked.R
\name{read_parts_chunked}
\alias{read_parts_chunked}
\title{read the fst parts in chunks
Note that the chunks may have a different size as the parts.}
\usage{
read_parts_chunked(x, chunk_size = 100000L, callback, ...)
}
\description{
read the fst parts in chunks
Note that the chunks may have a different size as the parts.
}
