% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repartition.R
\name{repartition_parts}
\alias{repartition_parts}
\title{resave the fstparts with specified chunk size and compression}
\usage{
repartition_parts(x, dir, chunksize = 1e+06, compress = 0, ...)
}
\description{
resave the fstparts with specified chunk size and compression
}
