% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewrite.R
\name{rewrite_parts}
\alias{rewrite_parts}
\title{resave the fstparts with specified chunk size and compression}
\usage{
rewrite_parts(x, path, chunksize = 1e+06, compress = 0)
}
\description{
resave the fstparts with specified chunk size and compression
}
