% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_part.R
\name{append_part}
\alias{append_part}
\title{Append data to a fstparts directory}
\usage{
append_part(x, data)
}
\arguments{
\item{x}{existing fstparts object.}

\item{data}{data.frame with data with same structure as \code{x}.}
}
\value{
updated fstparts object.
}
\description{
Apppend data to a fstparts directory
}
