% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstparts-package.r, R/fstparts.R
\docType{package}
\name{fstparts}
\alias{fstparts}
\alias{fstparts-package}
\alias{fstparts}
\title{fstparts}
\usage{
fstparts(data, dir = NULL, name = deparse(substitute(data)),
  overwrite = FALSE, compress = 0)
}
\description{
Store a large data set in chunks on disk in fst format, and make it easy to import, process and export the results.

fst
}
